core-parse-opts: minor code clean up

+++ core-parse-opts.c
@@ -572,7 +572,7 @@ uint64_t stress_get_uint64_percent(
 			longjmp(g_error_env, 1);
 			stress_no_return();
 		}
-		percent = ((double)max * (double)val) / (100.0 * (double)instances);
+		percent = ((double)max * val) / (100.0 * (double)instances);
 		if (percentage)
 			*percentage = true;
 		if (percent > (double)UINT64_MAX) {
@@ -751,23 +751,23 @@ int stress_parse_opt(const char *stressor_name, const char *opt_arg, const stres
 	case TYPE_ID_UINT64_BYTES:
 		/* uint64 in bytes units, generic */
 		setting.u.uint64 = stress_get_uint64_byte(opt_arg);
-		stress_check_range_bytes(opt_name, (uint64_t)setting.u.uint64, min, max);
+		stress_check_range_bytes(opt_name, setting.u.uint64, min, max);
 		return stress_setting_set(stressor_name, opt_name, TYPE_ID_UINT64_BYTES, &setting.u.uint64);
 	case TYPE_ID_UINT64_BYTES_FS:
 		/* uint64 in file system bytes units */
 		setting.u.uint64 = stress_get_uint64_byte_filesystem(opt_arg, 1, &percentage);
 		if (percentage)
 			return stress_setting_set(stressor_name, opt_name, TYPE_ID_UINT64_BYTES_FS_PERCENT, &setting.u.uint64);
-		stress_check_range_bytes(opt_name, (uint64_t)setting.u.uint64, min, max);
+		stress_check_range_bytes(opt_name, setting.u.uint64, min, max);
 		return stress_setting_set(stressor_name, opt_name, TYPE_ID_UINT64_BYTES_FS, &setting.u.uint64);
 	case TYPE_ID_UINT64_BYTES_VM:
 		/* uint64 in virtual memory bytes units */
 		setting.u.uint64 = stress_get_uint64_byte_memory(opt_arg, 1);
-		stress_check_range_bytes(opt_name, (uint64_t)setting.u.uint64, min, max);
+		stress_check_range_bytes(opt_name, setting.u.uint64, min, max);
 		return stress_setting_set(stressor_name, opt_name, TYPE_ID_UINT64_BYTES_VM, &setting.u.uint64);
 	case TYPE_ID_INT64:
 		setting.u.int64 = stress_get_int64(opt_arg);
-		stress_check_signed_range(opt_name, (int64_t)setting.u.int64, (int64_t)min, (int64_t)max);
+		stress_check_signed_range(opt_name, setting.u.int64, (int64_t)min, (int64_t)max);
 		return stress_setting_set(stressor_name, opt_name, TYPE_ID_INT64, &setting.u.int64);
 	case TYPE_ID_SIZE_T:
 		setting.u.size = (size_t)stress_get_uint64(opt_arg);
